Leaderboard completion
Added d3-request library to read csv fles
Updated CSS file for the podium
Added two components.
Leadership podium is the logic and code to have a leaderboard after reading the csv file
Ranking.js is the logic and code to have a table including teams 4-10 or all teams depending on if the podium will not show (condition is if more than 3 teams have the same points on a podium.